Making KATT happy
