--- components/pages/eventtypes/CreateFirstEventTypeView.tsx
@@ -1,30 +1,0 @@
-import { useLocale } from "@lib/hooks/useLocale";
-import { inferQueryOutput } from "@lib/trpc";
-
-import UserCalendarIllustration from "@components/ui/svg/UserCalendarIllustration";
-
-import CreateNewEventButton from "./CreateNewEventButton";
-
-type Profiles = inferQueryOutput<"viewer.eventTypes">["profiles"];
-
-interface Props {
-  canAddEvents: boolean;
-  profiles: Profiles;
-}
-
-const CreateFirstEventTypeView = ({ canAddEvents, profiles }: Props) => {
-  const { t } = useLocale();
-
-  return (
-    <div className="md:py-20">
-      <UserCalendarIllustration />
-      <div className="block mx-auto text-center md:max-w-screen-sm">
-        <h3 className="mt-2 text-xl font-bold text-neutral-900">{t("new_event_type_heading")}</h3>
-        <p className="mt-1 mb-2 text-md text-neutral-600">{t("new_event_type_description")}</p>
-        <CreateNewEventButton canAddEvents={canAddEvents} profiles={profiles} />
-      </div>
-    </div>
-  );
-};
-
-export default CreateFirstEventTypeView;
